Optimize FillLeft/FillRight
Go1.23 includes optimizations for "strings.Repeat". Use this function
results in better performance when filling space characters.
