+++ runewidth_fill.go
@@ -1,0 +1,25 @@
+//go:build go1.23
+
+package runewidth
+
+import "strings"
+
+// FillLeft return string filled in left by spaces in w cells
+func (c *Condition) FillLeft(s string, w int) string {
+	width := c.StringWidth(s)
+	count := w - width
+	if count > 0 {
+		return strings.Repeat(" ", count) + s
+	}
+	return s
+}
+
+// FillRight return string filled in left by spaces in w cells
+func (c *Condition) FillRight(s string, w int) string {
+	width := c.StringWidth(s)
+	count := w - width
+	if count > 0 {
+		return s + strings.Repeat(" ", count)
+	}
+	return s
+}
